Optimize test_python26 test based on log change
In adding the logging functionality, I modified the method
exit_if_unsupported_python so that instead of writing to stdout via
print, it wrote out to logging.critical. This broke the unit test
test_python26, as it was written to assert for equality of the first
line of that log output. I adjusted it so that it asserts on the exit
code of 1 instead.